requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
scipy>=1.10.0
scikit-learn>=1.2.0
fpdf2>=2.7.0
jupyter>=1.0.0
nbconvert>=7.0.0
kaleido>=0.2.1
streamlit>=1.25.0
